feat: Add endpoint to get all movie watch stats
This commit adds a new endpoint to the MovieWatchStatsController that returns all movie watch statistics. The new route is implemented with a GET request and uses the @ApiOperation and @ApiOkResponse decorators from Swagger. Additionally, the MovieWatchStatsService has been updated with a new method that retrieves all stats from the database.