feat(movie-watch-list): Add MovieHistoryService to providers
This commit adds the MovieHistoryService to the providers in the movie-watch-list module. It also updates the MovieWatchListService to use methods from the MovieHistoryService when deleting and updating watch list items. If a movie is marked as finished, it will update or create a history record for that user and movie.