feat: Add functionality to update episode/movie history
This commit adds the ability to update a user's movie or episode history when they finish watching an item. If the status of a watch list item is set to "FINISHED", the corresponding movie/episode history will be updated with a stoppedAt value of 1. If no previous history exists, a new entry will be created.