feat: Add endpoint to get history entry list
This commit adds a new endpoint to the `EpisodeHistoryController` that allows users to retrieve their viewing history for a list of media. The `getMediaHistoryList` method takes in an array of media IDs and returns the corresponding viewing history for each media. This feature was implemented by adding a new method called `getHistoryList` in the `EpisodeHistoryService`.